requirements.txt
discord.py>=1.7.3
python-dotenv>=0.21.0
Pillow>=9.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
